implement getting cities
Implemented the `getCitiesInGovernment` function in `LocationRepositoryImp` to fetch cities within a specified governorate and country. This involves creating a `CitiesRequest` and using the `locationRemoteDataSource` to make the API call.

Additionally, a `toCities` mapper function was added to `LocationMapper.kt` to convert the `CitiesDto` from the remote data source into a `Cities` domain entity.