Add requestedCount field to RequestServiceDto
The `requestedCount` field has been added to the `RequestServiceDto` data class. This field represents the number of times a service has been requested.

The `fromJson` companion object function has been updated to parse the `requestedCount` from the input data map. If the `requestedCount` is not present or is not a Long, it defaults to 0.