feat: Add RequestServiceUseCase
This commit introduces the `RequestServiceUseCase` class.

The `RequestServiceUseCase` class is responsible for handling the business logic related to requesting a service. It takes a `HomeRepository` as a dependency in its constructor.

The class has a single `invoke` operator function that takes a `RequestService` object as a parameter and calls the `requestService` method on the `HomeRepository`.