Refactor: Update onMarkAsDone to pass JobUiState object
The `onMarkAsDone` function in `MyJobCraftsmanInteractionListener` and its implementations now accepts a `JobUiState` object as a parameter instead of individual `requestId`, `requestTitle`, and `customerId` strings.

This change simplifies the function signature and allows for passing the entire job details when marking a job as done. Additionally, the `MyJobsCraftsmanViewModel` now uses `incrementJobsDoneForCraftsmanUseCase` and `updateEarningsForCraftsmanUseCase` within the `onMarkAsDone` function.